requirements.txt
# WhatsApp Integration
twilio==8.12.0

# AI/LLM
google-generativeai>=0.3.0
langchain-google-genai>=0.0.6
langchain>=0.1.10
langchain-core>=0.1.10

# Web Framework
flask==3.0.0
flask-cors>=4.0.0

# Data & Utilities
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0

# Jira Integration
atlassian-python-api>=3.41.0

# Optional: LangSmith for monitoring
langsmith>=0.0.77

# Additional dependencies your code uses
logging
typing-extensions>=4.5.0